+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 =========
 
 * Add support for CIF topologies in libmda using openmm.app.pdbxfile
+* removes the install*.yml gitactions
 
 v0.10.0 (2021-11-24)
 ------------------------------------------------------------